refactor: move ParallaxText to the common folder. Remove unused images.
