+++ src/components/Wallet/MobileAppParallax/index.tsx
@@ -1,4 +1,4 @@
-import ParallaxText, { type ParallaxTextProps } from '@/components/Wallet/ParallaxText'
+import ParallaxText, { type ParallaxTextProps } from '@/components/common/ParallaxText'
 import MobileAppParallaxElement from '@/components/Wallet/MobileAppParallax/MobileAppParallaxElement'
 
 const MobileAppParallax = (props: ParallaxTextProps) => (
