+++ src/components/Wallet/OwnershipParallax/index.tsx
@@ -1,4 +1,4 @@
-import ParallaxText, { type ParallaxTextProps } from '@/components/Wallet/ParallaxText'
+import ParallaxText, { type ParallaxTextProps } from '@/components/common/ParallaxText'
 import OwnershipParallaxElement from '@/components/Wallet/OwnershipParallax/OwnershipParallaxElement'
 
 const OwnershipParallax = (props: ParallaxTextProps) => (
